backend/ugc_service/ugc/requirements.txt
fastapi==0.93.0
grpcio-tools==1.51.3
gunicorn==20.1.0
uvicorn[standard]==0.20.0
pydantic==1.10.5
orjson==3.8.7
alembic==1.10.1
psycopg2-binary==2.9.5
email-validator==1.3.1
SQLAlchemy==2.0.4
SQLAlchemy-Utils==0.40.0
requests==2.28.2
fastapi-pagination==0.11.4
asyncpg==0.27.0
protobuf==4.22.1
backoff==2.2.1
celery==5.2.7
redis==4.5.1
